using i128 where unsigned ints(positive values) will need enough bits (greater than 64) to an i128

+++ src/lib.rs
@@ -164,21 +164,21 @@ impl<'a, I: Unit, T: Into<UnitArr<'a, I>>> Seek for BitCursor<'a, I, T> {
                 return Ok(seek_to);
             }
             SeekFrom::Current(v) => {
-                let unitsize = UnitArr::<I>::unit_size() as i64;
-                let bits = self.bit_pos as i64 + ((self.cur_position() as i64) * unitsize);
-                self.bit_pos = ((v + bits) % (unitsize)) as u8;
-                let seek_to = (v + bits) / unitsize;
+                let unitsize = UnitArr::<I>::unit_size() as i128;
+                let bits = self.bit_pos as i128 + ((self.cur_position() as i128) * unitsize);
+                self.bit_pos = (((v as i128) + bits) % (unitsize)) as u8;
+                let seek_to = ((v as i128) + bits) / unitsize;
                 self.set_cur_pos(seek_to as u64);
                 return Ok(seek_to as u64);
             }
             SeekFrom::End(v) => {
-                let unitsize = UnitArr::<I>::unit_size() as i64;
+                let unitsize = UnitArr::<I>::unit_size() as i128;
                 self.bit_pos = if v >= 0 {
-                    (v % unitsize) as u8
+                    ((v as i128) % unitsize) as u8
                 } else {
-                    (v % unitsize).wrapping_neg() as u8
+                    ((v as i128) % unitsize).wrapping_neg() as u8
                 };
-                let seek_to = v / unitsize;
+                let seek_to = (v as i128) / unitsize;
                 (self.cursor.get_ref().len() as u64, seek_to)
             }
         };
